[#30] Rename: popularPost 관련 컴포넌트 popularPosts 폴더로 이동
