+++ src/components/Board/Board.tsx
@@ -7,7 +7,7 @@ import { headers } from 'next/headers';
 import { BoardCategory } from '@/types/board';
 import { fetchPosts } from '@/service/firebase/firebase-firestore';
 import Posts from '@/components/Board/post/Posts';
-import PopularPosts from '@/components/Board/PopularPosts';
+import PopularPosts from '@/components/Board/popularPosts/PopularPosts';
 import WriteButton from '@/components/Board/write/WriteButton';
 
 type Props = { page: string | undefined };
